Create IOGatewayLBURLs.peoplecode
This plugin allows you to set the `LOCAL` gateway to be load balanced. Each PIA domain that runs this plugin will add itself to the list of Load Balanced IB Gateways, and you use use this plugin to set the main IB Gateway URL to the LB URL, instead of the local machine.